+++ CHANGELOG.md
@@ -4,12 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-##[3.0.1] 2018-11-14
+
+## [3.2.0] 2018-11-16
+
+### Added
+- Added getHex() to Actor.
+- Added getHex() getTransportType() and removePawn() a to Transport.
+
+### Fixed
+- Improved the documentation of IGameRunner
+- Improved the documentation of IGameBoard
+- GameEngine's constructor no longer leaks Common::GameException
+- GameEngine no longer tries to add boats to Hexes that are not water hexes.
+
+## [3.0.1] 2018-11-14
 
 ### Fixed
 - Added missing include directives.
 
-##[3.0.0] 2018-11-13
+## [3.0.0] 2018-11-13
 
 ### Changed
 - IGameBoard has been expanded, update your implementation.